BE-#13 refactor : review를 참고하여 코드 수정
- Entitiy 객체 Entity 클래스로 변경
