+++ BE/src/test/java/kr/codesquad/issuetracker/domain/entity/LabelTest.java
@@ -1,4 +1,4 @@
-package kr.codesquad.issuetracker.domain;
+package kr.codesquad.issuetracker.domain.entity;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -16,7 +16,7 @@ class LabelTest {
     @Test
     @DisplayName("NoArgsConstructor annotation test")
     void noArgsConstructorAnnotationTest() {
-        Label label = new Label();
+        Label label = Label.builder().build();
         assertThat(label.getId()).isNull();
         assertThat(label.getColor()).isNull();
         assertThat(label.getTitle()).isNull();
